Removes coroutines launcher abstraction
Although this seems like a good idea it is not needed and it can cause
more harm than good as other people won't be familiar with it and the
new keywords introduced by it.

Since unit tests are not affected by the absence non of this it is
better to remove it completely and use the native coroutines syntax.